move http_json into the testutils module
